codes/team19/relationformer-hack/requirements.txt
loguru==0.5.3
numpy
sknw==0.14
scikit_image==0.17.2
monai==0.7.0
matplotlib==3.2.1
pyvista==0.40
scipy==1.5.4
open3d
MedPy==0.4.0
mmcv_full==1.3.15
timm==0.4.12
tqdm==4.62.3
einops==0.3.2
mmcv==1.6.0
PyYAML==6.0
scikit_image==0.17.2
pytorch-ignite
tensorboard
webdataset